Completely re-write the system call routing system
With this change, system call meta-data is generated much more efficiently, and many aspects related to system calls that used to be arch-specific are now generic.
The main change implemented is to split the definition and binding of system calls, as well as remove all arch-specific code from the system call generator script.

+++ kos/src/libc/user/execinfo.c
@@ -271,7 +271,7 @@ NOTHROW_NCX(LIBCCALL libc_backtrace_symbols_fd)(void *const *array,
 		PRIVATE SECTION_DEBUG_STRING("debug_unknown_name") char const debug_unknown_name[1] = { '?' };
 		error = print_function_name(array[i],
 		                            (pformatprinter)(void *)&write,
-		                            (void *)(unsigned int)fd);
+		                            (void *)(uintptr_t)(unsigned int)fd);
 		if unlikely(error < 0)
 			break;
 		if (!error)
